Customize annotation and detail view
Added another view for the trees details that contain tree name,
scientific name, description and an image.fix the segue to pass in the
whole annotation object instead of its individual attributes.
Linked the annotation properties  to the labels in detail view.
Over wrote the annotation class to have an image and a button.
When the user presses on the annotation it show more details, when the
user presses on the I button it will segue to the detail view carrying
the specific tapped object and parsed it in the needed labels to
represent it.
Added an animation to the menu button.